move config metadata tests to a separate file
