+++ apps/studio/src/common/IPlatformInfo.ts
@@ -1,4 +1,5 @@
 import type { BksVersion } from '@/lib/license';
+import type { Platform } from '@/types';
 
 export interface IPlatformInfo {
   isWindows: boolean,
@@ -23,7 +24,7 @@ export interface IPlatformInfo {
   },
   debugEnabled: boolean,
   DEBUG: string,
-  platform: "linux" | "windows" | "mac",
+  platform: Platform,
   darkMode: boolean,
   userDirectory: string,
   downloadsDirectory: string,
